refactor: consolidate scroll tests into app_test.go
Move all scroll functionality tests from scroll_test.go into app_test.go
to follow Go naming conventions. The tests verify:
- Preview and tabbed window scroll methods
- Key binding mappings for scroll functions
- Mouse wheel scroll event handling
- Instance scroll position isolation
- Auto-scroll behavior and instance switching
- Fast scroll functionality (Ctrl+Shift+↑/↓)
- Duplicate title collision prevention
